Backend/requirements.txt
Flask
flask-cors
pydantic
langchain
langchain-google-genai
google-generativeai
waitress
manim
python-dotenv
